+++ R/show_sddf_cntrounds.R
@@ -24,10 +24,10 @@
 #' \dontrun{
 #' set_email("your_email@email.com")
 #' 
-#' show_sddf_rounds("Spain")
+#' show_sddf_cntrounds("Spain")
 #' }
 #'
-show_sddf_rounds <- function(country, ess_email = NULL) {
+show_sddf_cntrounds <- function(country, ess_email = NULL) {
   check_country(country)
   
   # Returns the chosen countries html that contains
@@ -61,8 +61,8 @@ show_sddf_rounds <- function(country, ess_email = NULL) {
   url_download <- grab_url_sddf_late_rounds(late_rounds, format = NULL)
 
   # Here I thought I might've introduced a bug because
-  # if I run show_sddf_rounds("Spain") and then
-  # show_sddf_rounds("Italy") then I thought that
+  # if I run show_sddf_cntrounds("Spain") and then
+  # show_sddf_cntrounds("Italy") then I thought that
   # because the sddf_laterounds_dir was already created and existed
   # for Spain, it WON'T download the data files for Italy. The thing
   # is that it doesn't matter because late rounds are integrated for