+++ NEWS.md
@@ -1,5 +1,9 @@
 ## essurvey 1.0.3.9999
 
+### Breaking changes
+
+* If you don't know which format is available for a round/country, `import_*` and `download_*` functions now accept a NULL argument which runs through `'stata'`, `'spss'` and `'sas'` formats automatically. By default, `import_*` functions have now format set to `NULL` to automatically try the three different formats. This breaks backward dependency but only slightly where it had 'stata' set as default.
+
 ### Bug fixes
 
 * Removes an unnecessary if statement in `set_email` that didn't allow to overwrite the email once set.