+++ NEWS.md
@@ -6,14 +6,16 @@
 
 ### New features
 
-* Users can now download SDDF (weight data) for each country/round combination of files. Functions `import_sddf_country` and `download_sddf_country` are now introduced. These functions only work for rounds `1:6` for now [https://github.com/ropensci/essurvey/commit/adb1da287e96de96713ce2cad9b4ed723b7ecf97].
+* Users can now download SDDF (weight data) for each country/round combination of files. Functions `import_sddf_country` and `download_sddf_country` are now introduced. These functions only work for rounds `1:6` for now [#9].
+
+
 
 ### Minor changes
 
 * Bumps `haven` to minimum package version 2.1.1
 
 ### Internal
-* `read_format_data` now tries to read data using `haven` but falls backs to `foreign` in case there's an error. This should only work for SDDF data [https://github.com/ropensci/essurvey/pulls?q=is%3Apr+is%3Aclosed].
+* `read_format_data` now tries to read data using `haven` but falls backs to `foreign` in case there's an error. This should only work for SDDF data [#38].
 
 ### Bug fixes